requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
django-ckeditor==6.5.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-js-asset==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
pep8-naming==0.13.3
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
